* Nueva opcion para Quitar Todos los Translados del Mapa * La funcion de deshacer ahora es activable/desactivable. * Agregada la opcion de Bloquear Arboles, Yacimientos y Carteles en el panel de Optimizar.

+++ Codigo/modDeclaraciones.bas
@@ -262,7 +262,7 @@ Public MaxYBorder As Byte
 Public Const MAX_INVENORY_OBJS  As Integer = 10000
 
 ' Deshacer
-Public Const maxDeshacer As Byte = 10
+Public Const maxDeshacer As Integer = 10
 Public MapData_Deshacer(1 To maxDeshacer, XMinMapSize To XMaxMapSize, YMinMapSize To YMaxMapSize) As MapBlock
 Type tDeshacerInfo
     Libre As Boolean
